+++ tikv-client/src/main/java/com/pingcap/tikv/meta/TiIndexInfo.java
@@ -42,6 +42,7 @@ public class TiIndexInfo implements Serializable {
   private final String comment;
   private final IndexType indexType;
   private final boolean isFakePrimaryKey;
+  private final boolean isInvisible;
 
   // default index column size (TypeFlag + Int64)
   private long indexColumnSize = 9;
@@ -60,7 +61,8 @@ public class TiIndexInfo implements Serializable {
       @JsonProperty("index_type") int indexType,
       // This is a fake property and added JsonProperty only to
       // to bypass Jackson frameworks's check
-      @JsonProperty("___isFakePrimaryKey") boolean isFakePrimaryKey) {
+      @JsonProperty("___isFakePrimaryKey") boolean isFakePrimaryKey,
+      @JsonProperty("is_invisible") boolean isInvisible) {
     this.id = id;
     this.name = requireNonNull(name, "index name is null").getL();
     this.tableName = requireNonNull(tableName, "table name is null").getL();
@@ -71,6 +73,7 @@ public class TiIndexInfo implements Serializable {
     this.comment = comment;
     this.indexType = IndexType.fromValue(indexType);
     this.isFakePrimaryKey = isFakePrimaryKey;
+    this.isInvisible = isInvisible;
   }
 
   public static TiIndexInfo generateFakePrimaryKeyIndex(TiTableInfo table) {
@@ -86,7 +89,8 @@ public class TiIndexInfo implements Serializable {
           SchemaState.StatePublic.getStateCode(),
           "Fake Column",
           IndexType.IndexTypeHash.getTypeCode(),
-          true);
+          true,
+          false);
     }
     return null;
   }
@@ -181,6 +185,10 @@ public class TiIndexInfo implements Serializable {
 
   public boolean isFakePrimaryKey() {
     return isFakePrimaryKey;
+  }
+
+  public boolean isInvisible() {
+    return isInvisible;
   }
 
   @Override